# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/Locations": {
            "get": {
                "description": "list Location entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List Location entities",
                "operationId": "list-Location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Location"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Location",
                "operationId": "create-Location",
                "parameters": [
                    {
                        "description": "Location entity",
                        "name": "Location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Location"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Location"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/Locations/{id}": {
            "get": {
                "description": "get Location by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a Location entity by ID",
                "operationId": "get-Location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Location"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update Location by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a Location entity by ID",
                "operationId": "update-Location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Location entity",
                        "name": "Location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Location"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Location"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get Location by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a Location entity by ID",
                "operationId": "delete-Location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ent.Activity": {
            "type": "object",
            "properties": {
                "CLUBE_ACTIVITY_NAME": {
                    "description": "CLUBEACTIVITYNAME holds the value of the \"CLUBE_ACTIVITY_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ActivityQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ActivityEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ActivityEdges": {
            "type": "object",
            "properties": {
                "club": {
                    "description": "Club holds the value of the club edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Club"
                    }
                }
            }
        },
        "ent.Club": {
            "type": "object",
            "properties": {
                "CLUB_NAME": {
                    "description": "CLUBNAME holds the value of the \"CLUB_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ClubQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ClubEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ClubEdges": {
            "type": "object",
            "properties": {
                "activity": {
                    "description": "Activity holds the value of the activity edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Activity"
                },
                "clubtypes": {
                    "description": "Clubtypes holds the value of the clubtypes edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ClubTypes"
                },
                "location": {
                    "description": "Location holds the value of the location edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Location"
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.ClubTypes": {
            "type": "object",
            "properties": {
                "CLUBE_TYPE_NAME": {
                    "description": "CLUBETYPENAME holds the value of the \"CLUBE_TYPE_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ClubTypesQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ClubTypesEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ClubTypesEdges": {
            "type": "object",
            "properties": {
                "club": {
                    "description": "Club holds the value of the club edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Club"
                    }
                }
            }
        },
        "ent.Location": {
            "type": "object",
            "properties": {
                "CLUBE_LOCATION_ADDRESS": {
                    "description": "CLUBELOCATIONADDRESS holds the value of the \"CLUBE_LOCATION_ADDRESS\" field.",
                    "type": "string"
                },
                "CLUBE_LOCATION_NAME": {
                    "description": "CLUBELOCATIONNAME holds the value of the \"CLUBE_LOCATION_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the LocationQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LocationEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.LocationEdges": {
            "type": "object",
            "properties": {
                "club": {
                    "description": "Club holds the value of the club edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Club"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "USER_EMAIL": {
                    "description": "USEREMAIL holds the value of the \"USER_EMAIL\" field.",
                    "type": "string"
                },
                "USER_NAME": {
                    "description": "USERNAME holds the value of the \"USER_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "club": {
                    "description": "Club holds the value of the club edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Club"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}